+++ jsonargparse/_common.py
@@ -100,6 +100,7 @@ parsing_settings = {
     "validate_defaults": False,
     "parse_optionals_as_positionals": False,
     "stubs_resolver_allow_py_files": False,
+    "omegaconf_absolute_to_relative_paths": False,
 }
 
 
@@ -112,6 +113,7 @@ def set_parsing_settings(
     docstring_parse_attribute_docstrings: Optional[bool] = None,
     parse_optionals_as_positionals: Optional[bool] = None,
     stubs_resolver_allow_py_files: Optional[bool] = None,
+    omegaconf_absolute_to_relative_paths: Optional[bool] = None,
 ) -> None:
     """
     Modify settings that affect the parsing behavior.
@@ -136,6 +138,10 @@ def set_parsing_settings(
             parser. By default, this is False.
         stubs_resolver_allow_py_files: Whether the stubs resolver should search
             in ``.py`` files in addition to ``.pyi`` files.
+        omegaconf_absolute_to_relative_paths: If True, when loading configs
+            with ``omegaconf+`` parser mode, absolute interpolation paths are
+            converted to relative. This is only intended for backward
+            compatibility with ``omegaconf`` parser mode.
     """
     # validate_defaults
     if isinstance(validate_defaults, bool):
@@ -162,6 +168,13 @@ def set_parsing_settings(
         parsing_settings["stubs_resolver_allow_py_files"] = stubs_resolver_allow_py_files
     elif stubs_resolver_allow_py_files is not None:
         raise ValueError(f"stubs_resolver_allow_py_files must be a boolean, but got {stubs_resolver_allow_py_files}.")
+    # omegaconf_absolute_to_relative_paths
+    if isinstance(omegaconf_absolute_to_relative_paths, bool):
+        parsing_settings["omegaconf_absolute_to_relative_paths"] = omegaconf_absolute_to_relative_paths
+    elif omegaconf_absolute_to_relative_paths is not None:
+        raise ValueError(
+            f"omegaconf_absolute_to_relative_paths must be a boolean, but got {omegaconf_absolute_to_relative_paths}."
+        )
 
 
 def get_parsing_setting(name: str):